+++ CHANGELOG.md
@@ -30,6 +30,8 @@ Thank you to all who have contributed!
 
 ### Added
 
+- Adds support for using EXCLUDED within DML ON-CONFLICT-ACTION conditions. Closes #1111.
+
 ### Changed
 
 - Updates Kotlin target from 1.4 (DEPRECATED) to 1.6
@@ -40,6 +42,10 @@ Thank you to all who have contributed!
   The file `partiql.ion` is still published in the `partiql-lang-kotlin` JAR.
 - Moves internal class org.partiql.lang.syntax.PartiQLParser to org.partiql.lang.syntax.impl.PartiQLPigParser as we refactor for explicit API.
 - Moves ANTLR grammar to `partiql-parser` package. The files `PartiQL.g4` and `PartiQLTokens.g4` are still published in the `partiql-lang-kotlin` JAR.
+- **Breaking**: Adds new property, `rowAlias`, to experimental `PartiqlLogical.DmlOperation.DmlUpdate`,
+  `PartiqlLogical.DmlOperation.DmlReplace`, `PartiqlLogicalResolved.DmlOperation.DmlUpdate`,
+  `PartiqlLogicalResolved.DmlOperation.DmlReplace`, `PartiqlPhysical.DmlOperation.DmlUpdate`, and
+  `PartiqlPhysical.DmlOperation.DmlReplace`.
 
 ### Deprecated